requirements.txt
streamlit>=1.35.0
plotly>=5.22.0
folium>=0.17.0
streamlit-folium>=0.22.0
opencv-python-headless>=4.10.0
Pillow>=10.4.0
numpy>=2.0.0
pandas>=2.2.0
fpdf2>=2.8.0
scikit-learn>=1.5.0
requests>=2.31.0
